requirements.txt
Flask
flask-cors
PyJWT
SQLAlchemy